requirements.txt
praw>=7.7.0
flask>=3.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
python-dotenv>=1.0.0